Project single cell gene expression data to developing mouse brain data (LaManno & Siletti et al. 2020)

Calculates correlation between single-cell gene expression and clusters from LaManno & Siletti et al. 2020

mousebrain_map(object, ...)

# S3 method for default
mousebrain_map(
  object,
  groups = NULL,
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = FALSE
)

# S3 method for Seurat
mousebrain_map(
  object,
  group_name = NULL,
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = FALSE
)

Arguments

groups
A character or factor vector or for grouping of cells, e.g. clusters, cell types.

method
A character string indicating which correlation coefficient to compute.

genes_use
A character vector with genes to use for computing the correlation. We recommend to use 150 - 500 genes.

allow_neg
Logical. Whether to allow negative correlations or set them to 0.

pseudobulk_groups
Logical. Whether to summarizse the group expression before computing the correlation.

group_name
A string indicating the metadata column for grouping the cells, e.g. clusters, cell types.

Value

A MousebrainMap object with a cell x ref correlation matrix and metadata.